+++ lib/data/blocs/auth/auth_bloc.dart
@@ -53,6 +53,12 @@ class AuthBloc extends Bloc<AuthEvent, AuthState> {
     if (event is LoginGuest) {
       yield* _mapGuestToState(event);
     }
+    if (event is LoginGoogle) {
+      yield* _mapGoogleToState(event);
+    }
+    if (event is ChangeUser) {
+      yield LoggedInState(event.user);
+    }
   }
 
   /// Called every time the user info changes. You can use this method for updating a database.
@@ -60,6 +66,18 @@ class AuthBloc extends Bloc<AuthEvent, AuthState> {
 
   /// Called when the user logs out. You can use this method for updating a database.
   final Function() deleteUser;
+
+  Stream<AuthState> _mapGoogleToState(LoginGoogle event) async* {
+    yield AuthLoadingState();
+    final _user = await _auth.loginGoogle(
+        idToken: event.idToken, accessToken: event.accessToken);
+    if (_user != null) {
+      if (saveUser != null) saveUser(_user);
+      yield LoggedInState(_user);
+    } else {
+      yield LoggedOutState();
+    }
+  }
 
   Stream<AuthState> _mapGuestToState(LoginGuest event) async* {
     yield AuthLoadingState();
@@ -85,13 +103,17 @@ class AuthBloc extends Bloc<AuthEvent, AuthState> {
 
   Stream<AuthState> _mapCreateToState(CreateAccount event) async* {
     yield AuthLoadingState();
-    AuthUser _user = await _auth.createAccount(event.username, event.password,
-        displayName: event?.displayName, photoUrl: event?.photoUrl);
-    if (_user != null) {
-      if (saveUser != null) saveUser(_user);
-      yield LoggedInState(_user);
-    } else {
-      yield AuthErrorState('Error creating user!');
+    try {
+      AuthUser _user = await _auth.createAccount(event.username, event.password,
+          displayName: event?.displayName, photoUrl: event?.photoUrl);
+      if (_user != null) {
+        if (saveUser != null) saveUser(_user);
+        yield LoggedInState(_user);
+      } else {
+        yield AuthErrorState('Error creating user!');
+      }
+    } catch (e) {
+      yield AuthErrorState('Email already exists!');
     }
   }
 
@@ -141,7 +163,7 @@ class AuthBloc extends Bloc<AuthEvent, AuthState> {
 
   static AuthUser currentUser(BuildContext context) {
     final auth = BlocProvider.of<AuthBloc>(context);
-    final state = auth.currentState;
+    final state = auth.state;
     if (state is LoggedInState) {
       return state.user;
     }