+++ lib/data/services/auth/web.dart
@@ -1,22 +1,24 @@
 import 'package:firebase/firebase.dart';
 
 import '../../classes/index.dart';
+import 'impl.dart';
 
-class FBAuth {
+class FBAuth implements FBAuthImpl {
   final FbApp app;
   final _auth = auth();
 
   FBAuth(this.app);
 
   Future _setPersistenceWeb(Auth _auth) async {
-    try {
-      var selectedPersistence = 'local';
-      await _auth.setPersistence(selectedPersistence);
-    } catch (e) {
-      print('_auth.setPersistence -> $e');
-    }
+    // try {
+    //   var selectedPersistence = 'local';
+    //   await _auth.setPersistence(selectedPersistence);
+    // } catch (e) {
+    //   print('_auth.setPersistence -> $e');
+    // }
   }
 
+  @override
   Future<AuthUser> login(String username, String password) async {
     await _setPersistenceWeb(_auth);
     try {
@@ -29,6 +31,7 @@ class FBAuth {
           email: _result.user?.email,
           isAnonymous: _result.user.isAnonymous,
           isEmailVerified: _result.user.emailVerified,
+          photoUrl: _result.user.photoURL,
         );
         return _user;
       }
@@ -38,6 +41,7 @@ class FBAuth {
     return null;
   }
 
+  @override
   Future<AuthUser> startAsGuest() async {
     await _setPersistenceWeb(_auth);
     try {
@@ -49,6 +53,7 @@ class FBAuth {
           email: _result.user?.email,
           isAnonymous: _result.user.isAnonymous,
           isEmailVerified: _result.user.emailVerified,
+          photoUrl: _result.user.photoURL,
         );
         return _user;
       }
@@ -58,6 +63,7 @@ class FBAuth {
     return null;
   }
 
+  @override
   Future logout() async {
     try {
       await _auth.signOut();
@@ -67,6 +73,7 @@ class FBAuth {
     return null;
   }
 
+  @override
   Stream<AuthUser> onAuthChanged() {
     return _auth.onAuthStateChanged.map((user) {
       if (user == null) return null;
@@ -76,11 +83,13 @@ class FBAuth {
         email: user?.email,
         isAnonymous: user.isAnonymous,
         isEmailVerified: user.emailVerified,
+        photoUrl: user.photoURL,
       );
       return _user;
     });
   }
 
+  @override
   Future<AuthUser> currentUser() async {
     await _setPersistenceWeb(_auth);
     try {
@@ -92,6 +101,7 @@ class FBAuth {
           email: _result?.email,
           isAnonymous: _result.isAnonymous,
           isEmailVerified: _result.emailVerified,
+          photoUrl: _result.photoURL,
         );
         return _user;
       }
@@ -101,6 +111,7 @@ class FBAuth {
     return null;
   }
 
+  @override
   Future editInfo({String displayName, String photoUrl}) async {
     final _user = _auth.currentUser;
     final _info = UserProfile();
@@ -113,6 +124,7 @@ class FBAuth {
     }
   }
 
+  @override
   Future forgotPassword(String email) async {
     try {
       await _auth.sendPasswordResetEmail(email);
@@ -121,6 +133,7 @@ class FBAuth {
     }
   }
 
+  @override
   Future sendEmailVerification() async {
     try {
       final _user = _auth.currentUser;
@@ -130,13 +143,68 @@ class FBAuth {
     }
   }
 
+  @override
   Future<AuthUser> createAccount(String username, String password,
       {String displayName, String photoUrl}) async {
-    final _user =
-        await _auth.createUserWithEmailAndPassword(username, password);
-    if (_user != null) {
-      await editInfo(displayName: displayName, photoUrl: photoUrl);
+    UserCredential _user;
+    try {
+      _user = await _auth.createUserWithEmailAndPassword(username, password);
+      if (_user != null) {
+        await editInfo(displayName: displayName, photoUrl: photoUrl);
+      }
+    } catch (e) {}
+    if (_user == null) {
+      try {
+        _user = await _auth.signInWithEmailAndPassword(username, password);
+      } catch (err) {
+        throw Exception(err);
+      }
     }
     return await currentUser();
+  }
+
+  @override
+  Future loginCustomToken(String token) async {
+    await _setPersistenceWeb(_auth);
+    try {
+      final _result = await _auth.signInAndRetrieveDataWithCustomToken(token);
+      if (_result != null && _result?.user != null) {
+        final _user = AuthUser(
+          uid: _result.user.uid,
+          displayName: _result.user.displayName,
+          email: _result.user?.email,
+          isAnonymous: _result.user.isAnonymous,
+          isEmailVerified: _result.user.emailVerified,
+          photoUrl: _result.user.photoURL,
+        );
+        return _user;
+      }
+    } catch (e) {
+      print('FBAuthUtils -> loginCustomToken -> $e');
+    }
+    return null;
+  }
+
+  @override
+  Future loginGoogle({String idToken, String accessToken}) async {
+    final _cred = GoogleAuthProvider.credential(idToken, accessToken);
+    await _setPersistenceWeb(_auth);
+    try {
+      final _result = await _auth.signInWithCredential(_cred);
+      if (_result != null && _result?.user != null) {
+        final _user = AuthUser(
+          uid: _result.user.uid,
+          displayName: _result.user.displayName,
+          email: _result.user?.email,
+          isAnonymous: _result.user.isAnonymous,
+          isEmailVerified: _result.user.emailVerified,
+          photoUrl: _result.user.photoURL,
+        );
+        return _user;
+      }
+    } catch (e) {
+      print('FBAuthUtils -> loginCustomToken -> $e');
+    }
+    return null;
   }
 }