+++ lib/data/services/mobile/sdk.dart
@@ -1,11 +1,10 @@
 import 'package:firebase_auth/firebase_auth.dart';
 
 import '../../classes/index.dart';
-import '../auth/unsupported.dart';
+import '../auth/impl.dart';
 
-class FbSdk implements FBAuth {
+class FbSdk implements FBAuthImpl {
   final _auth = FirebaseAuth.instance;
-  
   @override
   Future<AuthUser> login(String username, String password) async {
     try {
@@ -18,6 +17,7 @@ class FbSdk implements FBAuth {
           email: _result.user?.email,
           isAnonymous: _result.user.isAnonymous,
           isEmailVerified: _result.user.isEmailVerified,
+          photoUrl: _result.user.photoUrl,
         );
         return _user;
       }
@@ -37,6 +37,7 @@ class FbSdk implements FBAuth {
         email: user?.email,
         isAnonymous: user.isAnonymous,
         isEmailVerified: user.isEmailVerified,
+        photoUrl: user.photoUrl,
       );
       return _user;
     });
@@ -52,6 +53,7 @@ class FbSdk implements FBAuth {
         email: _result.user?.email,
         isAnonymous: _result.user.isAnonymous,
         isEmailVerified: _result.user.isEmailVerified,
+        photoUrl: _result.user.photoUrl,
       );
       return _user;
     }
@@ -78,6 +80,7 @@ class FbSdk implements FBAuth {
           email: _result?.email,
           isAnonymous: _result.isAnonymous,
           isEmailVerified: _result.isEmailVerified,
+          photoUrl: _result.photoUrl,
         );
         return _user;
       }
@@ -122,14 +125,73 @@ class FbSdk implements FBAuth {
   @override
   Future<AuthUser> createAccount(String username, String password,
       {String displayName, String photoUrl}) async {
-    final _user = await _auth.createUserWithEmailAndPassword(
-        email: username, password: password);
-    if (_user != null) {
-      await editInfo(displayName: displayName, photoUrl: photoUrl);
+    AuthResult _user;
+    try {
+      _user = await _auth.createUserWithEmailAndPassword(
+        email: username,
+        password: password,
+      );
+      if (_user != null) {
+        await editInfo(displayName: displayName, photoUrl: photoUrl);
+      }
+    } catch (e) {}
+    if (_user == null) {
+      try {
+        _user = await _auth.signInWithEmailAndPassword(
+          email: username,
+          password: password,
+        );
+      } catch (err) {
+        throw Exception(err);
+      }
     }
     return await currentUser();
   }
 
   @override
   FbApp get app => null;
+
+  @override
+  Future loginGoogle({String idToken, String accessToken}) async {
+    final _cred = GoogleAuthProvider.getCredential(
+        idToken: idToken, accessToken: accessToken);
+    try {
+      final _result = await _auth.signInWithCredential(_cred);
+      if (_result != null && _result?.user != null) {
+        final _user = AuthUser(
+          uid: _result.user.uid,
+          displayName: _result.user.displayName,
+          email: _result.user?.email,
+          isAnonymous: _result.user.isAnonymous,
+          isEmailVerified: _result.user.isEmailVerified,
+          photoUrl: _result.user.photoUrl,
+        );
+        return _user;
+      }
+    } catch (e) {
+      print('FBAuthUtils -> loginGoogle -> $e');
+    }
+    return null;
+  }
+
+  @override
+  Future loginCustomToken(String token) async {
+    try {
+      final _result = await _auth.signInWithCustomToken(token: token);
+      if (_result != null && _result?.user != null) {
+        final _user = AuthUser(
+          uid: _result.user.uid,
+          displayName: _result.user.displayName,
+          email: _result.user?.email,
+          isAnonymous: _result.user.isAnonymous,
+          isEmailVerified: _result.user.isEmailVerified,
+          photoUrl: _result.user.photoUrl,
+        );
+        return _user;
+      }
+    } catch (e) {
+      print('FBAuthUtils -> loginCustomToken -> $e');
+    }
+    return null;
+  }
 }