+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 1.1.0
+
+* Updating to Bloc 2.0.0
+
+## 1.0.6
+
+* Fixing Google Login (Mobile, Web)
+
+## 1.0.3
+
+* Updating to `bloc 1.0.0`
+* Fixing Google Login (Mobile, Web)
+* Try login after failed create account or user is null
+
+## 1.0.1
+
+* Support for Login with Access Token
+* Support for Google Sign In for Mobile and Web
+
 ## 1.0.0
 
 * Adding Support for Desktop